added a portfolio class
